refactor : articleResponseUtils 추가
